+++ packages/eslint-config-mints/rules/variables.js
@@ -48,7 +48,7 @@ module.exports = {
 
     // disallow unused variables
     // https://eslint.org/docs/rules/no-unused-vars
-    'no-unused-vars': ['error', {
+    'no-unused-vars': ['warn', {
       vars: 'all',
       args: 'after-used',
       ignoreRestSiblings: true