+++ packages/eslint-config-mints/CHANGELOG.md
@@ -1,3 +1,16 @@
+## 1.1.0
+
+`2019-07-18`
+
+* Update parser to babel-eslint.
+* Modify the ecmaVersion version to 2018.
+* Add option root to true.
+* Add some values about the options env and globals.
+
+Rule Change:
+
+  * `no-unused-vars`: change from error to warn.
+
 ## 1.0.0
 
 `2019-04-19`